pyproject.toml
[tool.poetry]
name = "inventory-app"
version = "0.1.0"
description = ""
authors = ["Henry Zheng <zhenghenry2@gmail.com>"]
readme = "README.md"
packages = [{include = "inventory_app"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
djangorestframework = "^3.14.0"
pygments = "^2.15.1"
django-extensions = "^3.2.3"
pandas = "^2.0.3"
django-cors-headers = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
